+++ app/controllers/admin/sections_controller.rb
@@ -1,2 +1,8 @@
 class Admin::SectionsController < Admin::ApplicationController
+  def import
+    agency = requested_resource
+    ImportCmsNodesJob.perform_later(agency.id)
+    flash.keep[:notice] = "Import scheduled for #{agency.name}."
+    redirect_to :action => :index
+  end
 end